Add linting to python dir.
- Fix oso
- Remove unused extras
