+++ languages/python/django-oso/setup.cfg
