+++ languages/python/oso/oso/__init__.py
@@ -1,3 +1,3 @@
-from polar import polar_class, Variable, Predicate
-from .oso import Oso, OsoError
-from .extras import Http, PathMapper
+from polar import polar_class, Variable, Predicate  # noqa: F401
+from .oso import Oso  # noqa: F401
+from polar.exceptions import OsoError  # noqa: F401
