+++ languages/python/oso/oso/oso.py
@@ -2,18 +2,7 @@
 
 __version__ = "0.8.2"
 
-from pathlib import Path
-import os
-from typing import Any, cast, Callable, List, TYPE_CHECKING
-import inspect
-
 from polar import Polar
-
-from .extras import Http, PathMapper
-from polar.exceptions import OsoError
-
-if TYPE_CHECKING:
-    import flask
 
 
 class Oso(Polar):
@@ -28,10 +17,6 @@ class Oso(Polar):
     def __init__(self):
         """Create an oso object."""
         super().__init__()
-
-        # Register built-in classes.
-        self.register_class(Http)
-        self.register_class(PathMapper)
 
     def is_allowed(self, actor, action, resource) -> bool:
         """Evaluate whether ``actor`` is allowed to perform ``action`` on ``resource``.
