+++ languages/python/oso/oso/test_oso.py
@@ -1,17 +1,10 @@
 """Tests the Polar API as an external consumer"""
 
-from authlib.jose import jwt
-from contextlib import contextmanager
-from flask import Flask, request, Response, g
 from pathlib import Path
-import pprint
 import pytest
-from dataclasses import dataclass
 
 from oso import Oso, polar_class
-from oso.jwt import Jwt
-from polar import Polar, Predicate
-from polar.test_helpers import public_key, private_key
+from polar import Predicate
 
 # Fake global actor name → company ID map.
 # Should be an external database lookup.
@@ -43,12 +36,7 @@ class Widget:
         return Company(id=self.id)
 
 
-@dataclass
 class Company:
-    # Data fields.
-    id: str = ""
-    default_role: str = ""
-
     # Class variables.
     roles = ("guest", "admin")
 
@@ -62,11 +50,13 @@ class Company:
         else:
             return "guest"
 
+    def __eq__(self, other):
+        return self.id == other.id
 
-@pytest.fixture(scope="module")
+
+@pytest.fixture
 def test_oso():
     oso = Oso()
-    oso.register_class(Jwt)
     oso.register_class(Actor)
     oso.register_class(Widget)
     oso.register_class(Company)
@@ -96,13 +86,6 @@ class Bar(Foo):
     bar: int = 1
 
 
-def token(name):
-    header = {"alg": "RS256"}
-    payload = {"iss": "somebody", "sub": name}
-    token = jwt.encode(header, payload, private_key)
-    return token.decode("utf-8")
-
-
 def test_is_allowed(test_oso):
     actor = Actor(name="guest")
     resource = Widget(id="1")
@@ -110,15 +93,12 @@ def test_is_allowed(test_oso):
     assert test_oso.is_allowed(actor, action, resource)
     assert test_oso.is_allowed({"username": "guest"}, action, resource)
     assert test_oso.is_allowed("guest", action, resource)
-    Jwt.add_key(public_key)
-    assert test_oso.is_allowed(token("guest"), action, resource)
 
     actor = Actor(name="president")
     action = "create"
     resource = Company(id="1")
     assert test_oso.is_allowed(actor, action, resource)
     assert test_oso.is_allowed({"username": "president"}, action, resource)
-    assert test_oso.is_allowed(token("president"), action, resource)
 
 
 def test_query_rule(test_oso):
@@ -134,8 +114,6 @@ def test_fail(test_oso):
     action = "not_allowed"
     assert not test_oso.is_allowed(actor, action, resource)
     assert not test_oso.is_allowed({"username": "guest"}, action, resource)
-    Jwt.add_key(public_key)
-    assert not test_oso.is_allowed(token("guest"), action, resource)
 
 
 def test_instance_from_external_call(test_oso):
@@ -143,8 +121,6 @@ def test_instance_from_external_call(test_oso):
     resource = Company(id="1")
     assert test_oso.is_allowed(user, "frob", resource)
     assert test_oso.is_allowed({"username": "guest"}, "frob", resource)
-    Jwt.add_key(public_key)
-    assert test_oso.is_allowed(token("guest"), "frob", resource)
 
 
 def test_allow_model(test_oso):
