+++ languages/python/oso/polar/__init__.py
@@ -1,9 +1,6 @@
-import json
-from pathlib import Path
-
-from .polar import Polar, polar_class
-from .query import Query, QueryResult
-from .variable import Variable
-from .predicate import Predicate
-from .partial import Partial
-from .expression import Expression, Pattern
+from .polar import Polar, polar_class  # noqa: F401
+from .query import Query, QueryResult  # noqa: F401
+from .variable import Variable  # noqa: F401
+from .predicate import Predicate  # noqa: F401
+from .partial import Partial  # noqa: F401
+from .expression import Expression, Pattern  # noqa: F401
