+++ languages/python/oso/requirements-test.txt
@@ -1,6 +1,2 @@
 pytest==5.3.5
 dataclasses ; python_version < '3.7'
-
-# Needed for flask, jwt extras
-authlib
-flask
