+++ languages/python/oso/requirements.txt
@@ -1,2 +1,1 @@
 cffi==1.14.0
-dataclasses ; python_version < '3.7'
