+++ languages/python/oso/setup.py
@@ -1,6 +1,5 @@
 import codecs
 import os
-import sys
 from setuptools import setup, find_packages
 from os import path
