+++ languages/python/requirements-lint.txt
@@ -1,0 +1,2 @@
+flake8
+mypy
